+++ data/src/message.rs
@@ -249,6 +249,11 @@ pub enum Target {
         channel: target::Channel,
         source: Source,
     },
+    ChannelMonitor {
+        server: Server,
+        channel: target::Channel,
+        source: Source,
+    },
 }
 
 impl Target {
@@ -265,6 +270,7 @@ impl Target {
             Target::Query { .. } => None,
             Target::Logs { .. } => None,
             Target::Highlights { .. } => None,
+            Target::ChannelMonitor { .. } => None,
         }
     }
 
@@ -275,6 +281,7 @@ impl Target {
             Target::Query { source, .. } => source,
             Target::Logs { source } => source,
             Target::Highlights { source, .. } => source,
+            Target::ChannelMonitor { source, .. } => source,
         }
     }
 
@@ -285,6 +292,7 @@ impl Target {
             Target::Query { source, .. } => source,
             Target::Logs { source } => source,
             Target::Highlights { source, .. } => source,
+            Target::ChannelMonitor { source, .. } => source,
         }
     }
 
@@ -294,7 +302,8 @@ impl Target {
             Target::Query { query, .. } => Some(query.as_str()),
             Target::Server { .. }
             | Target::Logs { .. }
-            | Target::Highlights { .. } => None,
+            | Target::Highlights { .. }
+            | Target::ChannelMonitor { .. } => None,
         }
     }
 
@@ -304,7 +313,8 @@ impl Target {
             Target::Query { .. }
             | Target::Server { .. }
             | Target::Logs { .. }
-            | Target::Highlights { .. } => None,
+            | Target::Highlights { .. }
+            | Target::ChannelMonitor { .. } => None,
         }
     }
 }
@@ -1069,6 +1079,13 @@ pub fn condense(
             Target::Highlights {
                 server, channel, ..
             } => Target::Highlights {
+                server: server.clone(),
+                channel: channel.clone(),
+                source,
+            },
+            Target::ChannelMonitor {
+                server, channel, ..
+            } => Target::ChannelMonitor {
                 server: server.clone(),
                 channel: channel.clone(),
                 source,